Add files via upload
first go at pulling historic data - only keep the last 31 days or so in the api. on investigating further found the historic stuff is dumped in csvs so need to come up with some new logic to just cycle through and pull the station i need out of the csv per day. wanted to save the logic for looping through a set of months though, hence still committing the file seperately